# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AirGo前后分离,多用户代理面板",
        "title": "AirGo",
        "contact": {
            "url": "https://github.com/pura-panel/airgo-panel"
        },
        "license": {
            "name": "GPL v3.0",
            "url": "https://github.com/pura-panel/airgo-panel/blob/main/LICENSE"
        }
    },
    "paths": {
...
        "/api/admin/menu/delMenu": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[admin api] menu"
                ],
                "summary": "删除菜单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Menu"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/admin/menu/getAllMenuList": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[admin api] menu"
                ],
                "summary": "获取全部菜单列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/admin/menu/newMenu": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[admin api] menu"
                ],
                "summary": "新建菜单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Menu"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
        "/api/admin/menu/updateMenu": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[admin api] menu"
                ],
                "summary": "修改菜单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户token",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Menu"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功；正常：业务代码 code=0；错误：业务代码code=1",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseStruct"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.AlipayPreCreatePayToFrontend": {
            "type": "object",
            "properties": {
                "qr_code": {
                    "type": "string"
                }
            }
        },
...
        "model.CasbinItem": {
            "type": "object",
            "properties": {
                "method": {
                    "description": "方法",
                    "type": "string"
                },
                "path": {
                    "description": "路径",
                    "type": "string"
                }
            }
        },
...
        "model.EpayPreCreatePay": {
            "type": "object",
            "properties": {
                "money": {
                    "description": "商品金额,如：1.00\t单位：元，最大2位小数",
                    "type": "string"
                },
                "name": {
                    "description": "商品名称,如超过127个字节会自动截取",
                    "type": "string"
                },
                "notify_url": {
                    "description": "服务器异步通知地址",
                    "type": "string"
                },
                "out_trade_no": {
                    "description": "商户订单号",
                    "type": "string"
                },
                "pid": {
                    "description": "商户ID",
                    "type": "integer"
                },
                "return_url": {
                    "description": "页面跳转通知地址",
                    "type": "string"
                },
                "sign": {
                    "description": "签名字符串，所有参数按照参数名ASCII码从小到大排序（a-z），sign、sign_type、和空值不参与签名！sign = md5 ( a=b\u0026c=d\u0026e=f + KEY ) （注意：+ 为各语言的拼接符，不是字符！），md5结果为小写。",
                    "type": "string"
                },
                "sign_type": {
                    "description": "签名类型",
                    "type": "string"
                },
                "type": {
                    "description": "支付方式",
                    "type": "string"
                }
            }
        },
        "model.EpayPreCreatePayToFrontend": {
            "type": "object",
            "properties": {
                "epay_api_url": {
                    "type": "string"
                },
                "epay_pre_create_pay": {
                    "$ref": "#/definitions/model.EpayPreCreatePay"
                }
            }
        },
...
        "model.Menu": {
            "type": "object",
            "properties": {
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Menu"
                    }
                },
                "component": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "meta": {
                    "$ref": "#/definitions/model.Meta"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "remarks": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Role"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "model.Meta": {
            "type": "object",
            "properties": {
                "icon": {
                    "type": "string"
                },
                "isAffix": {
                    "type": "boolean"
                },
                "isHide": {
                    "type": "boolean"
                },
                "isIframe": {
                    "type": "boolean"
                },
                "isKeepAlive": {
                    "type": "boolean"
                },
                "isLink": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "model.Order": {
            "type": "object",
            "properties": {
                "balance_amount": {
                    "type": "string"
                },
                "buyer_logon_id": {
                    "type": "string"
                },
                "buyer_pay_amount": {
                    "type": "string"
                },
                "coupon_amount": {
                    "type": "string"
                },
                "coupon_id": {
                    "type": "integer"
                },
                "coupon_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "customer_service_id": {
                    "description": "服务参数",
                    "type": "integer"
                },
                "deliver_text": {
                    "type": "string"
                },
                "deliver_type": {
                    "type": "string"
                },
                "des": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "goods_id": {
                    "description": "商品参数",
                    "type": "integer"
                },
                "goods_type": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "order_remarks": {
                    "type": "string"
                },
                "order_type": {
                    "type": "string"
                },
                "original_amount": {
                    "type": "string"
                },
                "out_trade_no": {
                    "type": "string"
                },
                "pay_id": {
                    "description": "支付参数",
                    "type": "integer"
                },
                "pay_info": {
                    "description": "支付信息，epay，alipay等\"",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.PreCreatePayToFrontend"
                        }
                    ]
                },
                "pay_type": {
                    "type": "string"
                },
                "price": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "string"
                },
                "trade_no": {
                    "type": "string"
                },
                "trade_status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "description": "关联用户",
                    "type": "integer"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
...
        "model.PreCreatePayToFrontend": {
            "type": "object",
            "properties": {
                "alipay_info": {
                    "$ref": "#/definitions/model.AlipayPreCreatePayToFrontend"
                },
                "epay_info": {
                    "$ref": "#/definitions/model.EpayPreCreatePayToFrontend"
                }
            }
        },
...
        "model.Role": {
            "type": "object",
            "properties": {
                "casbins": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CasbinItem"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "menus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Menu"
                    }
                },
                "role_name": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_group": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.User"
                    }
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "balance": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "enable": {
                    "type": "boolean"
                },
                "enable_email": {
                    "type": "boolean"
                },
                "enable_tg_bot": {
                    "description": "通知参数",
                    "type": "boolean"
                },
                "enable_web_mail": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "invitation_code": {
                    "type": "string"
                },
                "nick_name": {
                    "type": "string"
                },
                "orders": {
                    "description": "订单，has many",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Order"
                    }
                },
                "password": {
                    "type": "string"
                },
                "referrer_user_id": {
                    "type": "integer"
                },
                "role_group": {
                    "description": "关联参数",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Role"
                    }
                },
                "tg_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                },
                "when_balance_changed": {
                    "type": "boolean"
                },
                "when_purchased": {
                    "type": "boolean"
                },
                "when_service_almost_expired": {
                    "type": "boolean"
                }
            }
        },
...
        "response.ResponseStruct": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}